+++ examples/nextjs-file-upload/minio/setup.sh
@@ -6,12 +6,12 @@ docker-compose up -d
 sleep 3
 
 function setupMinio () {
-    docker-compose exec -T minio mc alias set myminio/ http://127.0.0.1:9000 minio minio123
-    docker-compose exec -T minio mc admin user add myminio test 12345678
-    docker-compose exec -T minio mc admin policy set myminio readwrite user=test
-    docker-compose exec -T minio mc admin user info myminio test
-    docker-compose exec -T minio mc mb myminio/uploads
-    docker-compose exec -T minio mc policy set public myminio/uploads
+    docker-compose exec -T minio mc alias set minio http://localhost:9000 minio minio123
+    docker-compose exec -T minio mc admin user add minio test 12345678
+    docker-compose exec -T minio mc admin policy attach minio readwrite -u test
+    docker-compose exec -T minio mc admin user info minio test
+    docker-compose exec -T minio mc mb --ignore-existing minio/uploads
+    docker-compose exec -T minio mc anonymous set public minio/uploads
     return $?
 }