+++ packages/testsuite/apps/auth/.wundergraph/wundergraph.server.ts
@@ -15,27 +15,26 @@ export default configureWunderGraphServer(() => ({
 		queries: {},
 		mutations: {},
 		authentication: {
-			postAuthentication: async (hook) => {
-				console.log(`postAuthentication: ${JSON.stringify(hook.user)}`);
-			},
+			postAuthentication: async (hook) => {},
 			mutatingPostAuthentication: async (hook) => {
-				console.log(`mutatingPostAuthentication: ${JSON.stringify(hook.user)}`);
 				if (hook.user.userId === 'notAllowedByPostAuthentication') {
 					return {
 						status: 'deny',
 						message: 'denied by mutatingPostAuthentication',
 					};
 				}
+				const customClaims = hook.user.customClaims ?? {};
+				customClaims.aTestValue = ['anArray'];
 				return {
 					status: 'ok',
 					user: {
 						...hook.user,
+						customClaims,
 						firstName: hook.user.firstName ?? 'mutatingPostAuthentication',
 					},
 				};
 			},
 			revalidate: async (hook) => {
-				console.log(`revalidate: ${JSON.stringify(hook.user)}`);
 				if (hook.user.userId === 'notAllowedByRevalidation') {
 					return {
 						status: 'deny',
@@ -50,9 +49,7 @@ export default configureWunderGraphServer(() => ({
 					},
 				};
 			},
-			postLogout: async (hook) => {
-				console.log(`postLogout: ${JSON.stringify(hook.user)}`);
-			},
+			postLogout: async (hook) => {},
 		},
 	},
 	graphqlServers: [