+++ packages/testsuite/apps/basic/.wundergraph/operations/__underscores__/42_as_a_string.ts
@@ -1,0 +1,7 @@
+import { createOperation } from '../../generated/wundergraph.factory';
+
+export default createOperation.query({
+	handler: async () => {
+		return '42';
+	},
+});