+++ testapps/default/.wundergraph/wundergraph.config.ts
@@ -9,7 +9,6 @@ import {
 import server from './wundergraph.server';
 import operations from './wundergraph.operations';
 import generate from './wundergraph.generate';
-import { golangClient } from '@wundergraph/golang-client';
 
 const jsp = introspect.openApiV2({
 	id: 'jsp',
@@ -55,6 +54,14 @@ const federatedApi = introspect.federation({
 const spacex = introspect.graphql({
 	apiNamespace: 'spacex',
 	url: 'https://spacex-api.fly.dev/graphql/',
+	schemaExtension: `
+	extend type Capsule {
+		myCustomField: String
+	}
+	`,
+	introspection: {
+		disableCache: true,
+	},
 });
 
 const countries = introspect.graphql({