+++ packages/svelte-query/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.20](https://github.com/wundergraph/wundergraph/compare/@wundergraph/svelte-query@0.2.19...@wundergraph/svelte-query@0.2.20) (2023-05-04)
+
+**Note:** Version bump only for package @wundergraph/svelte-query
+
+## [0.2.19](https://github.com/wundergraph/wundergraph/compare/@wundergraph/svelte-query@0.2.18...@wundergraph/svelte-query@0.2.19) (2023-05-03)
+
+**Note:** Version bump only for package @wundergraph/svelte-query
+
+## [0.2.18](https://github.com/wundergraph/wundergraph/compare/@wundergraph/svelte-query@0.2.17...@wundergraph/svelte-query@0.2.18) (2023-05-03)
+
+**Note:** Version bump only for package @wundergraph/svelte-query
+
+## [0.2.17](https://github.com/wundergraph/wundergraph/compare/@wundergraph/svelte-query@0.2.16...@wundergraph/svelte-query@0.2.17) (2023-04-27)
+
+**Note:** Version bump only for package @wundergraph/svelte-query
+
+## [0.2.16](https://github.com/wundergraph/wundergraph/compare/@wundergraph/svelte-query@0.2.15...@wundergraph/svelte-query@0.2.16) (2023-04-27)
+
+**Note:** Version bump only for package @wundergraph/svelte-query
+
+## [0.2.15](https://github.com/wundergraph/wundergraph/compare/@wundergraph/svelte-query@0.2.14...@wundergraph/svelte-query@0.2.15) (2023-04-27)
+
+**Note:** Version bump only for package @wundergraph/svelte-query
+
+## [0.2.14](https://github.com/wundergraph/wundergraph/compare/@wundergraph/svelte-query@0.2.13...@wundergraph/svelte-query@0.2.14) (2023-04-26)
+
+**Note:** Version bump only for package @wundergraph/svelte-query
+
+## [0.2.13](https://github.com/wundergraph/wundergraph/compare/@wundergraph/svelte-query@0.2.12...@wundergraph/svelte-query@0.2.13) (2023-04-26)
+
+**Note:** Version bump only for package @wundergraph/svelte-query
+
 ## [0.2.12](https://github.com/wundergraph/wundergraph/compare/@wundergraph/svelte-query@0.2.11...@wundergraph/svelte-query@0.2.12) (2023-04-25)
 
 **Note:** Version bump only for package @wundergraph/svelte-query